refactor: - move order data parsing to server side - remove order/orders event listeners
