+++ backend/server.js
@@ -5,7 +5,7 @@ const app = require("./app");
 const database = require("./src/database/db");
 const sessionMiddleware = require("./src/middleware/session");
 const config = require("./src/config");
-const { corsConfig, wrap } = require("./src/controllers/server.controller");
+const { corsConfig, wrap } = require("./src/utils/server.utils");
 const OrderingSession = require("./src/components/OrderingSession");
 
 const server = http.createServer(app);
